imfusion_sdk/requirements.txt
imfusion-sdk==0.8.0
matplotlib==3.9.0
jupyter==1.0.0
ipywidgets==8.1.5
